% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_matrix.R
\name{vector_to_matrix}
\alias{vector_to_matrix}
\title{Matrix Mean and Variance}
\usage{
vector_to_matrix(MCMC_dat, diffuse_se = 100, diffuse = TRUE, intercept0 = TRUE)
}
\arguments{
\item{MCMC_dat}{A data.table object containing MCMC samples.}

\item{diffuse_se}{A numeric value specifying the standard error for the diffuse prior.}

\item{diffuse}{A logical value indicating whether to use a diffuse prior. Defaults to \code{TRUE}.}

\item{intercept0}{A logical value indicating whether to center the intercept term of the meta-regression to zero. Defaults to \code{TRUE}.}
}
\value{
A list containing:
\describe{
\item{means}{A list of mean vectors for each MCMC sample.}
\item{covar}{A list of covariance matrices for each MCMC sample.}
\item{dt}{A data.table containing the mean and covariance values for each MCMC sample.}
}
}
\description{
Generate mean and covariance using prior data. We will use this to construct weights to update prior \eqn{\xi}.
}
\examples{
# Use the example MCMC data provided in the package `historical_posterior`
result_matrix<-vector_to_matrix(
MCMC_dat = historical_posterior,
diffuse_se = 100,
diffuse = TRUE,
intercept0 = TRUE
)

head(result_matrix)
}
