% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_posterior}
\alias{historical_posterior}
\title{Posterior Distribution of Parameter Values from Modeling with UACR, Acute Slope, and Chronic Slope as Surrogates for Clinical Endpoint}
\format{
A data.table with 500 rows and 4 variables:
\describe{
\item{col1}{Description}
\item{col2}{Description}
}
}
\usage{
historical_posterior
}
\description{
Posterior distribution of all parameters from the MCMC model with UACR,
acute slope, and chronic slope as surrogates for the clinical endpoint.
This data set is used in the examples of the package to demonstrate how
to use the \code{vector_to_matrix} and \code{MBE} functions.
}
\keyword{datasets}
