% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBE.R
\name{MBE}
\alias{MBE}
\title{Estimate posterior distribution of Multi-Component Bayesian Endpoints (MBE)}
\usage{
MBE(
  mcmc_dat,
  sample_dat,
  diffuse = TRUE,
  diffuse_se = 100,
  ncores = 3,
  intercept0 = TRUE
)
}
\arguments{
\item{mcmc_dat}{A data.table object containing MCMC samples.}

\item{sample_dat}{A list containing the new RCT data (estimated treatment effects on the CE and surrogates and their variances and correlations).
They should explicitely have names \code{ClnEst}, \code{ClnSE},  \code{Sur1Est}, \code{Sur1SE}, \code{Sur2Est}, \code{Sur2SE}, \code{R1Clin}, \code{R2Clin}, and \code{R12}. See Example.}

\item{diffuse}{A logical value indicating whether to use a diffuse prior. Defaults to TRUE.}

\item{diffuse_se}{A numeric value specifying the standard error for the diffuse prior.}

\item{ncores}{An integer. Number of parallel chains to run. Defaults to 3. If set to 1, the function will run sequentially without parallelization.}

\item{intercept0}{A logical value indicating whether to center the intercept term of the meta-regression to zero. Defaults to TRUE.
stripped before the computation proceeds. Defaults to \code{TRUE}.}
}
\value{
A data frame with three columns: \code{mean}, \code{sd}, and \code{n}.
}
\description{
Takes new data and historical posterior samples to estimate the posterior distribution of MBE using importance sampling.
Built specifically for the case of two surrogate endpoints and a single clinical endpoint.
}
\examples{
# Use the example data provided in the package.
# Load the historical posterior samples where the model was fit
# with random intercept and inverse gamma prior for variance parameters
# on the 65/66 simulated trials in `trial_sim_dat` dataset.
# It's the first trial that was left out as if it was a new trial.
# We demonstrate estimating the posterior distribution of MBE on this first
# row that was left out.
data("historical_posterior")
data("trial_sim_dat")
one_sim_dat<-list(

# treatment effect on CE
ClnEst=trial_sim_dat$CE_est[1],
ClnSE=trial_sim_dat$CE_se[1],

# treatment effect on chronic slope
Sur1Est=trial_sim_dat$Sur1_est[1],
Sur1SE=trial_sim_dat$Sur1_se[1],

# treatment effect on acute slope
Sur2Est=trial_sim_dat$Sur2_est[1],
Sur2SE=trial_sim_dat$Sur2_se[1],

# correlation between CE and chronic slope
R1Clin=trial_sim_dat$Cor_CE_Sur1[1],

# correlation between CE and acute slope
R2Clin=trial_sim_dat$Cor_CE_Sur2[1],

# correlation between chronic slope and acute slope
R12=trial_sim_dat$Cor_Sur1_Sur2[1]
)

MBE_distribution<-MBE(
mcmc_dat = historical_posterior,
sample_dat=one_sim_dat,
diffuse_se = 100,
diffuse = TRUE,
intercept0 = TRUE
)

head(MBE_distribution)
}
