% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBE.R
\name{MBE}
\alias{MBE}
\title{Estimate posterior distribution of Multi-Component Bayesian Endpoints (MBE)}
\usage{
MBE(
  mcmc_dat,
  sample_dat,
  diffuse = TRUE,
  diffuse_se = 100,
  ncores = 3,
  intercept0 = TRUE
)
}
\arguments{
\item{mcmc_dat}{A data.table object containing MCMC samples.}

\item{sample_dat}{A list containing the observed mean and variance vector.}

\item{diffuse}{A logical value indicating whether to use a diffuse prior. Defaults to TRUE.}

\item{diffuse_se}{A numeric value specifying the standard error for the diffuse prior.}

\item{ncores}{An integer. For parallel computing.}

\item{intercept0}{A logical value indicating whether to center the intercept term of the meta-regression to zero. Defaults to TRUE.
stripped before the computation proceeds. Defaults to \code{TRUE}.}
}
\value{
A data frame with three columns: \code{mean}, \code{sd}, and \code{n}.
}
\description{
Takes a numeric vector and returns a data frame containing
the mean, standard deviation, and count of non-missing values.
}
\examples{
# Use the example MCMC data provided in the package.
# Use a single time point in the provided simulated data.

# We will use simulation case 2, month 30.
sim_dat<-data.table::data.table(sim_dat)
sim_dat<-sim_dat[case==2 & analysis.month==30, ]
one_sim_dat<-list(

# treatment effect on CE
ClnEst=sim_dat$ClnEst,
ClnSE=sim_dat$ClnSE,

# treatment effect on chronic slope
Sur1Est=sim_dat$Sur1Est,
Sur1SE=sim_dat$Sur1SE,

# treatment effect on acute slope
Sur2Est=sim_dat$Sur2Est,
Sur2SE=sim_dat$Sur2SE,

# correlation between CE and chronic slope
R1Clin=sim_dat$R1Clin,

# correlation between CE and acute slope
R2Clin=sim_dat$R2Clin,

# correlation between chronic slope and acute slope
R1R2=sim_dat$R12
)

MBE_distribution<-MBE(
mcmc_dat = historical_posterior,
sample_dat=one_sim_dat,
diffuse_se = 100,
diffuse = TRUE,
intercept0 = TRUE
)

head(MBE_distribution)
}
