% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trial_sim_dat}
\alias{trial_sim_dat}
\title{66 Trial Level Summary Simulated Data}
\format{
A data frame with 66 rows (one row per trial) and 13 variables:
\describe{
\item{trial_id}{Integer. Index for trial.}
\item{CE_est}{Numeric. Observed estimated treatment effect on the clinical endpoint.}
\item{Sur1_est}{Numeric. Observed estimated treatment effect on the first surrogate endpoint.}
\item{Sur2_est}{Numeric. Observed estimated treatment effect on the second surrogate endpoint.}
\item{CE_se}{Numeric. Standard error of the observed estimated treatment effect on the clinical endpoint.}
\item{Sur1_se}{Numeric. Standard error of the observed estimated treatment effect on the first surrogate endpoint.}
\item{Sur2_se}{Numeric. Standard error of the observed estimated treatment effect on the second surrogate endpoint.}
\item{Cor_CE_Sur1}{Numeric. Estimated correlation between clinical endpoint and 1st surrogate endpoint.}
\item{Cor_CE_Sur2}{Numeric. Estimated correlation between clinical endpoint and 2nd surrogate endpoint.}
\item{Cor_Sur1_Sur2}{Numeric. Estimated correlation between 1st surrogate endpoint and 2nd surrogate endpoint.}
\item{theta_CE}{Numeric. True treatment effect on the clinical endpoint.}
\item{theta_Sur1}{Numeric. True treatment effect on the 1st surrogate endpoint.}
\item{theta_Sur2}{Numeric. True treatment effect on the 2nd surrogate endpoint.}
}
}
\source{
Simulated by Yizhen Xu
}
\usage{
trial_sim_dat
}
\description{
Wait until Yizhen gives me her description
}
\author{
Yizhen Xu \email{yizhen.xu@utah.edu}
}
\keyword{datasets}
